Add Opportunity data
I added a new data store for opportunities, and a new page that displays them.
I left the page intentionally unstyles and unformatted, because I'm not sure
what's going on in other branches to add a layout system and stuff.  It'll be
easy to add that stuff after the fact.